Add files via upload
I think this is actually simpler than i thought.
I thought that we had to have the ability to add multiple reservation to a single room as long as the dates overlapped.
But since we have to add a function that "outputs true if all room are reserved, i think only one reservation can be had on a room at any given time. What do you think?